+++ CHANGES.md
@@ -3,7 +3,8 @@
 
 ### Version 13.5.5, New development version, not released yet
   
-  * No changes yet
+  * Add minimum and maximum to Data.SBV.List. If they receive empty list as argument,
+    then the result is underspecified, i.e., can be any value of the element type.
 
 ### Version 13.5, 2026-01-26